Add preemptible arg for core and worker nodes
This change allows users to make core and/or worker node pools preemptible.
The commit also removes the "preemptible:true" taint. The thinking here is
that the worker node already has a "dedicated:worker" so users have
an existing mechanism to select the worker or core node pool without
the "preemptible" taint.